Don't clean non-existent directory

+++ Module.BuildSettings.ps1
@@ -290,7 +290,7 @@ Task AfterClean -After Clean {
     $BinPath = Join-Path $SrcRootDir "bin"
 
     # Maybe a bit paranoid but this task nuked \ on my laptop. Good thing I was not running as admin.
-    if ($BinPath.Length -gt 3) {
+    if ($BinPath.Length -gt 3 -and (Test-path $BinPath)) {
         Get-ChildItem $Binpath | Remove-Item -Recurse -Force -Verbose:$VerbosePreference
     } else {
         Write-Verbose "$($Task.Name) - `$BinPath '$BinPath' must be longer than 3 characters."
